requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
SQLAlchemy==2.0.40
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Pillow==11.2.1
Flask-Migrate==4.1.0
